requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-appconf==1.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-filter==24.3
django-parler==2.3
django-parler-rest==2.2
django-staticfiles==1.2.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
parler==1.0.1
pillow==10.4.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==73.0.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
